requirements.txt
discord-components-mirror==2.1.5
discord-webhook[async]==1.0.0
discord.py==1.7.3
discord==1.7.3
Flask[async]==2.2.2
fuzzywuzzy==0.18.0
Pillow==9.3.0
pymongo==4.3.3
PyNaCl==1.5.0
pypresence==4.2.1
pytz==2022.6
pywebio-battery==0.4.0
pywebio==1.7.1
requests==2.30.0
Werkzeug==2.2.2
